refactor: 현상관리 메인 페이지 리팩토링(#135)
