+++ src/pages/photoManage/PhotoManageMainPage.tsx
@@ -2,52 +2,23 @@ import { useEffect, useState } from "react";
 import { useQuery } from "@tanstack/react-query";
 import { useNavigate } from "react-router";
 import ProcessStep from "@/components/photoManage/ProcessStep";
-import {
-  DevelopPicIcon,
-  ScanPicIcon,
-  PrinterIcon,
-  PrintPicIcon,
-  DeliveryPicIcon,
-  MenuIcon,
-  DownloadIcon,
-  PencilLineIcon,
-  ClockIcon,
-  TruckIcon,
-  CheckEmptyIcon,
-  CircleIcon,
-} from "@/assets/icon";
+import { MenuIcon, CircleIcon } from "@/assets/icon";
 import { Header } from "@/components/common";
-import { ActionButton } from "@/components/photoManage/ActionButton";
-import { RecipientInfoCard } from "@/components/photoManage/RecipientInfoCard";
 import { DialogBox } from "@/components/common/DialogBox";
 import BottomSheet from "@/components/common/BottomSheet";
 import { getCurrentWork } from "@/apis/photoManage/currentWork.api";
-import type {
-  Status,
-  ReceiptMethod,
-  DevelopmentStatus,
-  // DeliveryStatus,
-  // MyCurrentWorkResponse as CurrentWorkData,
-} from "@/types/photomanage/process";
-
-type StepConfig = {
-  key: string;
-  step: Status;
-  receiptMethod?: ReceiptMethod;
-  isCurrent: boolean;
-  title: string;
-  content: React.ReactNode;
-  index: number;
-  subComment?: React.ReactNode;
-  buttons?: React.ReactNode;
-  isLast?: boolean;
-};
+import { type Status } from "@/types/photomanage/process";
+import { buildProcessSteps } from "@/lib/buildProcessSteps";
+import { getActiveStatus } from "@/utils/getActiveStatus";
+import { STATUS_INDEX_MAP } from "@/constants/photomanage/status.constant";
+import { getBannerContent } from "@/lib/getBannerContent";
 
 export default function PhotoManageMainPage() {
   const navigate = useNavigate();
   const [isDialogOpen, setIsDialogOpen] = useState(false);
   const [dialogStep, setDialogStep] = useState(1);
 
+  // 진행 상황 조회
   const {
     data: currentWorkResponse,
     isLoading,
@@ -65,231 +36,27 @@ export default function PhotoManageMainPage() {
     if (isError) navigate("/development-history");
   }, [isSuccess, isError, workData, navigate]);
 
-  const getActiveStatus = (): Status => {
-    if (!workData) return "DEVELOP";
-
-    if (workData.delivery && workData.delivery.deliveryId) return "DELIVERY";
-
-    if (workData.print && workData.print.printOrderId) return "PRINT";
-
-    const devStatusMap: Record<DevelopmentStatus, Status> = {
-      RECEIVED: "DEVELOP",
-      DEVELOPING: "DEVELOP",
-      SCANNING: "SCAN",
-      COMPLETED: "SCAN",
-    };
-
-    const currentStatus = devStatusMap[workData.developmentStatus] ?? "DEVELOP";
-
-    return currentStatus;
-  };
-  const status: Status = getActiveStatus();
+  const status: Status = getActiveStatus(workData);
   const receiptMethod = workData?.print?.receiptMethod;
-
-  const STATUS_INDEX_MAP: Record<Status, number> = {
-    DEVELOP: 1,
-    SCAN: 2,
-    PRINT: 3,
-    DELIVERY: 4,
-  };
   const currentIndex = STATUS_INDEX_MAP[status];
 
-  // --- 헬퍼 함수 ---
-  const formatEstimatedTime = (isoDate: string | null): string => {
-    if (!isoDate) return "현재 확인 중";
-    const date = new Date(isoDate);
-    const hour = date.getHours();
-    return `${date.getMonth() + 1}월 ${date.getDate()}일 ${hour >= 12 ? "오후" : "오전"} ${
-      hour > 12 ? hour - 12 : hour === 0 ? 12 : hour
-    }시`;
-  };
+  const currentBanner = getBannerContent({
+    status,
+    receiptMethod,
+    printStatus: workData?.print?.status,
+  });
 
-  const formatShippedDate = (isoDate: string | null): string => {
-    if (!isoDate) return "-";
-    const date = new Date(isoDate);
-    const days = ["일", "월", "화", "수", "목", "금", "토"];
-    return `${date.getFullYear()}.${String(date.getMonth() + 1).padStart(2, "0")}.${String(
-      date.getDate(),
-    ).padStart(2, "0")}(${days[date.getDay()]})`;
-  };
-
-  // --- UI 데이터 맵 ---
-  const getBannerContent = () => {
-    const bannerMap = {
-      DEVELOP: {
-        icon: <DevelopPicIcon />,
-        title: "필름 현상중이에요",
-        content: "현상이 완료되면 이곳에서 사진을 볼 수 있어요!",
-      },
-      SCAN: {
-        icon: <ScanPicIcon />,
-        title: "현상된 필름이 스캔 완료되었어요",
-        content: "인화 신청과 사진 다운로드를 해주세요!",
-      },
-      PRINT: {
-        icon: <PrintPicIcon />,
-        title:
-          workData?.print?.status === "PENDING"
-            ? "현상소에서 확인 중이에요"
-            : "인화 작업이 진행 중이에요",
-        content: "배송이 시작되면 알려드릴게요!",
-      },
-      DELIVERY: {
-        icon: <DeliveryPicIcon />,
-        title:
-          receiptMethod === "PICKUP"
-            ? "사진 수령이 가능해요"
-            : "사진을 배송하고 있어요",
-        content:
-          receiptMethod === "PICKUP"
-            ? "현상소에 방문하여 사진을 찾아주세요!"
-            : "현상소에서 발송된 사진이 이동 중이에요!",
-      },
-    };
-    return bannerMap[status];
-  };
-
-  const currentBanner = getBannerContent();
-
-  const steps: StepConfig[] = workData
-    ? [
-        {
-          key: "DEVELOP",
-          step: "DEVELOP",
-          isCurrent: status === "DEVELOP",
-          title: "필름 현상",
-          content:
-            status === "DEVELOP"
-              ? "아직 현상이 완료되지 않았어요"
-              : "어둠 속에서 이미지를 깨우는 중",
-          index: 1,
-        },
-        {
-          key: "SCAN",
-          step: "SCAN",
-          isCurrent: status === "SCAN",
-          title: "디지털 스캔",
-          content:
-            status === "SCAN"
-              ? "인화여부를 확정해야 다음 단계로 넘어가요!"
-              : "고해상도 디지털 파일로 변환",
-          buttons: status === "SCAN" && (
-            <div className="flex flex-col gap-2.5">
-              <ActionButton
-                leftIcon={<PrinterIcon />}
-                message="인화 여부 확정하기"
-                showNext
-                onClick={() => setIsDialogOpen(true)}
-              />
-              <ActionButton
-                leftIcon={<DownloadIcon />}
-                message="사진 다운로드 하러 가기"
-                showNext
-                onClick={() => navigate("/photoManage/download")}
-              />
-            </div>
-          ),
-          index: 2,
-        },
-        {
-          key: "PRINT",
-          step: "PRINT",
-          receiptMethod,
-          isCurrent: status === "PRINT",
-          title: "사진 인화",
-          subComment: status === "PRINT" && workData.print && (
-            <div>
-              <p className="mb-0.5 flex items-center gap-1 text-[0.8125rem] text-[#EC602D]">
-                <ClockIcon className="h-3 w-3" />
-                {workData.print.status === "PENDING"
-                  ? "작업 예정 시간 확인 중"
-                  : `예상 작업 완료 시간: ${formatEstimatedTime(workData.print.estimatedAt)}`}
-              </p>
-              <hr className="mb-1.5 border-orange-500/30" />
-            </div>
-          ),
-          content:
-            status === "PRINT" && workData.print?.status === "PRINTING"
-              ? "종이 위에 추억이 새겨지는 중..."
-              : "선명한 사진 프린트 작업",
-          buttons: status === "PRINT" && (
-            <ActionButton
-              leftIcon={<PencilLineIcon className="h-4 w-4" />}
-              message="인화하는 동안 사진 자랑하러 가기"
-              showNext
-              onClick={() => navigate("/photoFeed")}
-            />
-          ),
-          index: 3,
-        },
-        {
-          key: "DELIVERY",
-          step: "DELIVERY",
-          receiptMethod,
-          isCurrent: status === "DELIVERY",
-          title: receiptMethod === "PICKUP" ? "방문 수령" : "택배 배송",
-          subComment: status === "DELIVERY" && workData.delivery && (
-            <div>
-              <p className="mb-0.5 flex items-center gap-1 text-[0.8125rem] text-[#EC602D]">
-                {workData.delivery.status === "DELIVERED" ? (
-                  <CheckEmptyIcon className="h-3 w-3" />
-                ) : (
-                  <TruckIcon className="text-orange-450 h-3 w-3" />
-                )}
-                {workData.delivery.status === "DELIVERED"
-                  ? "배송 완료"
-                  : "배송 중"}
-              </p>
-              <hr className="mb-1.5 border-orange-500/30" />
-            </div>
-          ),
-          content:
-            status === "DELIVERY" && workData.delivery ? (
-              <RecipientInfoCard
-                items={[
-                  {
-                    label: "발송일",
-                    value: formatShippedDate(workData.delivery.shippedAt),
-                  },
-                  { label: "택배사", value: workData.delivery.carrier ?? "-" },
-                  {
-                    label: "송장 번호",
-                    value: workData.delivery.trackingNumber ?? "-",
-                    copyValue: workData.delivery.trackingNumber ?? undefined,
-                  },
-                ]}
-              />
-            ) : receiptMethod === "PICKUP" ? (
-              "현상소 방문 후 수령해 주세요"
-            ) : (
-              "안전하게 포장하여 배송"
-            ),
-          buttons: status === "DELIVERY" && (
-            <div className="flex flex-col gap-2.5">
-              {receiptMethod === "DELIVERY" && (
-                <ActionButton
-                  leftIcon={
-                    <TruckIcon className="text-orange-450 h-4.5 w-4.5" />
-                  }
-                  message="배송 조회하러 가기"
-                  showNext
-                  onClick={() => {}}
-                />
-              )}
-              {workData.delivery?.status === "DELIVERED" && (
-                <ActionButton
-                  leftIcon={<CheckEmptyIcon className="h-4 w-4" />}
-                  message="수령 확정 하기"
-                  showNext
-                  onClick={() => {}}
-                />
-              )}
-            </div>
-          ),
-          index: 4,
-          isLast: true,
-        },
-      ]
+  const steps = workData
+    ? buildProcessSteps({
+        workData,
+        status,
+        receiptMethod,
+        onOpenPrintConfirmDialog: () => setIsDialogOpen(true),
+        onGoDownload: () => navigate("/photoManage/download"),
+        onGoFeed: () => navigate("/photoFeed"),
+        onGoTrackDelivery: () => {},
+        onConfirmReceived: () => {},
+      })
     : [];
 
   if (isLoading)
@@ -360,6 +127,7 @@ export default function PhotoManageMainPage() {
           onClose={() => {}}
           sheetClassName="bg-neutral-900/60 backdrop-blur-3xl"
           isBackDrop={false}
+          collapsedRatio={0.7}
         >
           <div className="flex h-full flex-col gap-4 p-4">
             <p className="mb-1 flex justify-start text-[0.8125rem] text-white">
