+++ src/types/photomanage/process.ts
@@ -1,6 +1,11 @@
+// 작업 진행 단계
 export type Status = "DEVELOP" | "SCAN" | "PRINT" | "DELIVERY";
-export type SpecStep = "BEFORE" | "AFTER";
+
+// 수령 방식
 export type ReceiptMethod = "PICKUP" | "DELIVERY";
+
+// 배송 상태
+export type DeliveryStatus = "PENDING" | "SHIPPED" | "DELIVERED";
 
 // 현상 주문 상태
 export type DevelopmentStatus =
@@ -17,9 +22,6 @@ export type PrintOrderStatus =
   | "READY"
   | "SHIPPED"
   | "COMPLETED";
-
-// 배송 상태
-export type DeliveryStatus = "PENDING" | "SHIPPED" | "DELIVERED";
 
 // 인화 진행 상태
 export interface PrintProgressResponse {
@@ -54,124 +56,20 @@ export interface MyCurrentWorkResponse {
   photoLabName: string;
   developmentStatus: DevelopmentStatus;
   createdAt: string;
+  compltedAt: string;
   print: PrintProgressResponse | null;
   delivery: DeliveryProgressResponse | null;
 }
 
-export type Process = {
-  status: Status;
-  specStep?: SpecStep;
-  receiptMethod?: ReceiptMethod;
-  print?: PrintProgressResponse;
-  delivery?: DeliveryProgressResponse;
-  deliveryInfo?: DeliveryInfo;
+export type StepConfig = {
+  key: string;
+  step: Status; // 작업 진행 단계
+  receiptMethod?: ReceiptMethod; // 수령 방식
+  isCurrent: boolean; // 현재 진행 여부
+  title: string;
+  content: React.ReactNode;
+  index: number;
+  subComment?: React.ReactNode;
+  buttons?: React.ReactNode;
+  isLast?: boolean;
 };
-
-export const developMock: Process = {
-  status: "DEVELOP",
-};
-export const scanMock: Process = {
-  status: "SCAN",
-};
-export const printMock: Process = {
-  status: "PRINT",
-  specStep: "BEFORE",
-  receiptMethod: "DELIVERY",
-};
-export const deliveryMock: Process = {
-  status: "DELIVERY",
-  specStep: "BEFORE",
-  receiptMethod: "DELIVERY",
-  deliveryInfo: {
-    recipientName: "김필름",
-    recipientPhone: "010-1234-5678",
-    address: "서울 동작구 흑석로 123-123 (123호)",
-  },
-};
-
-// PM-070-1: 현상소 확인 중
-export const printPendingMock: Process = {
-  status: "PRINT",
-  specStep: "BEFORE",
-  receiptMethod: "DELIVERY",
-  print: {
-    printOrderId: 101,
-    status: "PENDING",
-    receiptMethod: "DELIVERY",
-    estimatedAt: null,
-    completedAt: null,
-  },
-  deliveryInfo: {
-    recipientName: "김필름",
-    recipientPhone: "010-1234-5678",
-    address: "서울 동작구 흑석로 123-123 (123호)",
-  },
-};
-
-// PM-070-2: 인화 작업 진행 중
-export const printConfirmedMock: Process = {
-  status: "PRINT",
-  specStep: "AFTER",
-  receiptMethod: "DELIVERY",
-  print: {
-    printOrderId: 101,
-    status: "PRINTING",
-    receiptMethod: "DELIVERY",
-    estimatedAt: "2026-01-17T15:00:00+09:00",
-    completedAt: null,
-  },
-  deliveryInfo: {
-    recipientName: "김필름",
-    recipientPhone: "010-1234-5678",
-    address: "서울 동작구 흑석로 123-123 (123호)",
-  },
-};
-
-// 배송중
-export const deliveryShippedMock: Process = {
-  status: "DELIVERY",
-  receiptMethod: "DELIVERY",
-  delivery: {
-    deliveryId: 1,
-    status: "SHIPPED",
-    carrier: "우체국 택배",
-    trackingNumber: "123412341234",
-    shippedAt: "2026-01-17T10:00:00+09:00",
-    deliveredAt: null,
-  },
-  deliveryInfo: {
-    recipientName: "파인더스 상도점",
-    recipientPhone: "010-1234-5678",
-    address: "서울 동작구 흑석로 123-123 (123호)",
-  },
-};
-
-// 배송 완료
-export const deliveryCompletedMock: Process = {
-  status: "DELIVERY",
-  receiptMethod: "DELIVERY",
-  delivery: {
-    deliveryId: 1,
-    status: "DELIVERED",
-    carrier: "우체국 택배",
-    trackingNumber: "123412341234",
-    shippedAt: "2026-01-17T10:00:00+09:00",
-    deliveredAt: "2026-01-19T14:30:00+09:00",
-  },
-  deliveryInfo: {
-    recipientName: "파인더스 상도점",
-    recipientPhone: "010-1234-5678",
-    address: "서울 동작구 흑석로 123-123 (123호)",
-  },
-};
-
-export const mocks = {
-  develop: developMock,
-  scan: scanMock,
-  print: printMock,
-  delivery: deliveryMock,
-  printPending: printPendingMock,
-  printConfirmed: printConfirmedMock,
-  deliveryShipped: deliveryShippedMock,
-  deliveryCompleted: deliveryCompletedMock,
-} as const;
